no se

+++ src/clases/Libro.cc
@@ -54,7 +54,7 @@ Libro::~Libro(){
     if (getNinio() != NULL){
     vector<Objeto*> nuevo_obj;
     vector<Objeto*> viejo_obj = getNinio()->getObjetosPrestados();
-    for(int i = 0; i < viejo_obj.size(); i++){
+    for(int i = 0; i < (int)viejo_obj.size(); i++){
         if (viejo_obj[i]->toString() != this->toString())
             nuevo_obj.push_back(viejo_obj[i]);
     }
